build: Debug|x64 builds and links for the first time; the build script gates on unresolved externals
Two build-infrastructure changes, verified by producing this tree's first
ever SwgClient_d.exe (the Debug x64 configuration had never compiled,
let alone linked).

THE GATE (scripts/Build-Client.ps1): SwgClient links with /FORCE, which
downgrades unresolved-external errors to warnings and emits a binary
anyway -- MSBuild exit 0 has never meant a clean link in this repo. The
script now tees the build output to a log and FAILS the build if any
'unresolved external symbol' line appears, printing the offenders. This
gate is what surfaced every defect below.

THE DEBUG REPAIR, outermost-in:

- Tag.h / MeshConstructionHelper.cpp / crypto filters.h: size_t
  narrowing and template-argument fixes for code that only compiles (or
  only warnings-as-errors) in Debug. clientAudio / clientDirectInput /
  crypto had TreatWarningAsError only in their Debug configs -- flipped
  to match their own working Release configs.
- libMozilla.vcxproj: Debug|x64 becomes ConfigurationType Utility --
  the same fix Release|x64 already carries, same reason (the legacy
  XULRunner headers cannot compile under modern x64 SDKs; the link uses
  the hand-written stub deps\x64\lib\libMozilla.lib).
- swg.sln: the dpvs Debug|x64 row mapped to an 'IntelCPP|x64' project
  configuration that no longer exists; now Debug|x64 -> Debug|x64.
- dpvsMemory.cpp: the Debug-only memory-guard tags cast pointers to
  UINT32 (C4311 under the x64 guardrail); routed through UPTR, low-32
  tag semantics unchanged -- same shape as the earlier dpvs x64 fixes.
- SwgClient.vcxproj Debug|x64: LinkIncremental off (incremental silently
  disabled the /FORCE this config sets -- LNK4075); the dependency list
  drops the legacy x86 third-party soup (Qt 3/4, soePlatform, maya,
  alienbrain, videocapture... -- x86 libs can never contribute to an x64
  link; with the soup gone the REAL unresolved set surfaced); dpvs and
  lcdui reference the solution-built Debug libs explicitly (the deps dir
  copies are Release/MT builds and outrank the build dir on the libpath
  -- they were 130 LNK2038 debug-CRT mismatches); dinput8+dxguid added
  (were only reachable via the soup).
- deps/x64: vivoxSharedWrapper_d.lib -- the wrapper rebuilt /MTd from
  the in-tree Vivox.cpp (compiles with VIVOX_VERSION=3, matching the
  in-tree SDK headers); videocapture-stubs_d.lib + its committed source
  (compat-source/videocapture_stubs.cpp) -- no-op definitions for the
  VideoCapture/SoeUtil symbols SwgVideoCapture.obj references. The real
  implementations only exist as x86 libs, so video capture has been
  structurally dead on x64 since the port; Release never notices because
  nothing there pulls that object out of clientGraphics.lib. The stubs
  make the Debug pull harmless (capture reports cStateDead).

Verify: Build-Client -Configuration Debug -Architecture x64 -Renderer
DX9 completes with zero unresolved externals; SwgClient_d.exe +
gl05/06/07_d produced. Boot smoke of the Debug client is the remaining
field check.

+++ src/engine/client/library/clientSkeletalAnimation/src/shared/appearance/MeshConstructionHelper.cpp
@@ -553,7 +553,7 @@ void MeshConstructionHelper::prepareForReading(void)
 			{
 				// get the VertexData for this shader vertex
 				const size_t vertexDataIndex = psd->m_vertexIndices[shaderVertexIndex];
-				VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0U, vertexDataIndex, m_vertexData->m_container.size()); //lint !e1703 // arbitrarily chosen template function
+				VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(static_cast<size_t>(0), vertexDataIndex, m_vertexData->m_container.size()); //lint !e1703 // arbitrarily chosen template function
 
 				VertexData &vertexData = m_vertexData->m_container[vertexDataIndex];
 
@@ -1402,7 +1402,7 @@ const MeshConstructionHelper::TriStripHeader &MeshConstructionHelper::getTriStri
 	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE( 0, triStripIndex, static_cast<int>(perShaderData->m_triStripIndices.size()) );
 
 	const size_t triStripHeaderIndex = perShaderData->m_triStripIndices[static_cast<size_t>(triStripIndex)];
-	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE( 0U, triStripHeaderIndex, m_triStripHeaders->m_container.size() ); //lint !e1703
+	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE( static_cast<size_t>(0), triStripHeaderIndex, m_triStripHeaders->m_container.size() ); //lint !e1703
 
 	return m_triStripHeaders->m_container[triStripHeaderIndex];
 }
@@ -1429,7 +1429,7 @@ int MeshConstructionHelper::getTriStripVertexIndex(const TriStripHeader &triStri
 	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0, vertexIndex, triStripHeader.m_indexCount);
 
 	const size_t indexIntoTriStripIndex = static_cast<size_t>(triStripHeader.m_firstTriStripIndex + vertexIndex);
-	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0U, indexIntoTriStripIndex, m_triStripIndices->m_container.size() ); //lint !e1703
+	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(static_cast<size_t>(0), indexIntoTriStripIndex, m_triStripIndices->m_container.size() ); //lint !e1703
 
 	return static_cast<int>(m_triStripIndices->m_container[indexIntoTriStripIndex]);
 }
@@ -1451,7 +1451,7 @@ const MeshConstructionHelper::TriListHeader &MeshConstructionHelper::getTriList(
 	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0, triListIndex, static_cast<int>(perShaderData->m_triListIndices.size()));
 
 	const size_t triListHeaderIndex = perShaderData->m_triListIndices[static_cast<size_t>(triListIndex)];
-	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0U, triListHeaderIndex, m_triListHeaders->m_container.size()); //lint !e1703 // template arbitrarily selected
+	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(static_cast<size_t>(0), triListHeaderIndex, m_triListHeaders->m_container.size()); //lint !e1703 // template arbitrarily selected
 
 	return m_triListHeaders->m_container[triListHeaderIndex];
 }
@@ -1472,7 +1472,7 @@ void MeshConstructionHelper::getTriListTriangle(const PerShaderData *perShaderDa
 	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0, triIndex, triListHeader.m_triangleCount);
 
 	const size_t indexedTriangleIndex = triListHeader.m_firstTriangleIndex + static_cast<size_t>(triIndex);
-	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(0U, indexedTriangleIndex, m_indexedTriangles->size() / 3); //lint !e1703 // template arbitrarily selected
+	VALIDATE_RANGE_INCLUSIVE_EXCLUSIVE(static_cast<size_t>(0), indexedTriangleIndex, m_indexedTriangles->size() / 3); //lint !e1703 // template arbitrarily selected
 
 	//const IndexedTriangle &triangle = (*m_indexedTriangles)[indexedTriangleIndex];
